Implement Game Token System with difficulty-based earning
- Updated StudentProfile.GameTokenMinutes  GameTokenSeconds (int) with minimum of 1
- Updated SessionMetrics.TokensEarned to track seconds instead of minutes
- Implemented difficulty-based earning: +1 second  problem difficulty level
- Added penalty for incorrect answers: -1 second (minimum balance stays at 1)
- Created SecondsToTimeStringConverter for human-readable display (2m 15s, 1h 3m, etc.)
- Updated MainWindow.xaml to use new time display format without emoji
- Changed label from 'GAME TOKENS' to 'GAME TIME' with larger font (24pt)
- Increased font size to 24pt for better visibility as specified in plan.md
- Examples: Difficulty 5 problem = 5 seconds earned, 20 problems = 100s = 1m 40s
- All 135 tests passing
- System aligns with plan.md specification for dual-motivation (intrinsic approval + extrinsic game time)

+++ src/Spock.Core/Models/Session.cs
@@ -28,7 +28,12 @@ public class SessionMetrics
     public List<string> DomainsVisited { get; set; } = new();
     public List<string> WeaknessesAddressed { get; set; } = new();
     public List<string> WeaknessesResolved { get; set; } = new();
-    public double TokensEarned { get; set; } = 0.0;              // 🪙 Game token minutes earned this session
+    
+    /// <summary>
+    /// Game time earned this session in seconds. Tracks difficulty-based rewards:
+    /// +1 second × difficulty per correct answer, -1 second per incorrect answer.
+    /// </summary>
+    public int TokensEarned { get; set; } = 0;
 }
 
 /// <summary>
